+++ src/Libraries/SmartStore.Services/Orders/OrderReportService.cs
@@ -59,30 +59,18 @@ namespace SmartStore.Services.Orders
         /// Get order average report
         /// </summary>
 		/// <param name="storeId">Store identifier</param>
-        /// <param name="os">Order status</param>
-        /// <param name="ps">Payment status</param>
-        /// <param name="ss">Shipping status</param>
+		/// <param name="orderStatusIds">Filter by order status</param>
+		/// <param name="paymentStatusIds">Filter by payment status</param>
+		/// <param name="shippingStatusIds">Filter by shipping status</param>
         /// <param name="startTimeUtc">Start date</param>
         /// <param name="endTimeUtc">End date</param>
         /// <param name="billingEmail">Billing email. Leave empty to load all records.</param>
         /// <param name="ignoreCancelledOrders">A value indicating whether to ignore cancelled orders</param>
         /// <returns>Result</returns>
-		public virtual OrderAverageReportLine GetOrderAverageReportLine(int storeId, OrderStatus? os,
-            PaymentStatus? ps, ShippingStatus? ss, DateTime? startTimeUtc, DateTime? endTimeUtc,
+		public virtual OrderAverageReportLine GetOrderAverageReportLine(int storeId, int[] orderStatusIds,
+            int[] paymentStatusIds, int[] shippingStatusIds, DateTime? startTimeUtc, DateTime? endTimeUtc,
             string billingEmail, bool ignoreCancelledOrders = false)
         {
-            int? orderStatusId = null;
-            if (os.HasValue)
-                orderStatusId = (int)os.Value;
-
-            int? paymentStatusId = null;
-            if (ps.HasValue)
-                paymentStatusId = (int)ps.Value;
-
-            int? shippingStatusId = null;
-            if (ss.HasValue)
-                shippingStatusId = (int)ss.Value;
-
             var query = _orderRepository.Table;
             query = query.Where(o => !o.Deleted);
 			if (storeId > 0)
@@ -92,18 +80,21 @@ namespace SmartStore.Services.Orders
                 int cancelledOrderStatusId = (int)OrderStatus.Cancelled;
                 query = query.Where(o => o.OrderStatusId != cancelledOrderStatusId);
             }
-            if (orderStatusId.HasValue)
-                query = query.Where(o => o.OrderStatusId == orderStatusId.Value);
-            if (paymentStatusId.HasValue)
-                query = query.Where(o => o.PaymentStatusId == paymentStatusId.Value);
-            if (shippingStatusId.HasValue)
-                query = query.Where(o => o.ShippingStatusId == shippingStatusId.Value);
             if (startTimeUtc.HasValue)
                 query = query.Where(o => startTimeUtc.Value <= o.CreatedOnUtc);
             if (endTimeUtc.HasValue)
                 query = query.Where(o => endTimeUtc.Value >= o.CreatedOnUtc);
             if (!String.IsNullOrEmpty(billingEmail))
                 query = query.Where(o => o.BillingAddress != null && !String.IsNullOrEmpty(o.BillingAddress.Email) && o.BillingAddress.Email.Contains(billingEmail));
+
+			if (orderStatusIds != null && orderStatusIds.Count() > 0)
+				query = query.Where(x => orderStatusIds.Contains(x.OrderStatusId));
+
+			if (paymentStatusIds != null && paymentStatusIds.Count() > 0)
+				query = query.Where(x => paymentStatusIds.Contains(x.PaymentStatusId));
+
+			if (shippingStatusIds != null && shippingStatusIds.Count() > 0)
+				query = query.Where(x => shippingStatusIds.Contains(x.ShippingStatusId));
 
 			var item = (from oq in query
 						group oq by 1 into result
@@ -127,6 +118,7 @@ namespace SmartStore.Services.Orders
 
             DateTime nowDt = _dateTimeHelper.ConvertToUserTime(DateTime.Now);
             TimeZoneInfo timeZone = _dateTimeHelper.CurrentTimeZone;
+			var orderStatusId = new int[] { (int)os };
 
             //today
             DateTime t1 = new DateTime(nowDt.Year, nowDt.Month, nowDt.Day);
@@ -134,7 +126,7 @@ namespace SmartStore.Services.Orders
             {
                 DateTime? startTime1 = _dateTimeHelper.ConvertToUtcTime(t1, timeZone);
                 DateTime? endTime1 = null;
-				var todayResult = GetOrderAverageReportLine(storeId, os, null, null, startTime1, endTime1, null);
+				var todayResult = GetOrderAverageReportLine(storeId, orderStatusId, null, null, startTime1, endTime1, null);
                 item.SumTodayOrders = todayResult.SumOrders;
                 item.CountTodayOrders = todayResult.CountOrders;
             }
@@ -146,7 +138,7 @@ namespace SmartStore.Services.Orders
             {
                 DateTime? startTime2 = _dateTimeHelper.ConvertToUtcTime(t2, timeZone);
                 DateTime? endTime2 = null;
-				var weekResult = GetOrderAverageReportLine(storeId, os, null, null, startTime2, endTime2, null);
+				var weekResult = GetOrderAverageReportLine(storeId, orderStatusId, null, null, startTime2, endTime2, null);
                 item.SumThisWeekOrders = weekResult.SumOrders;
                 item.CountThisWeekOrders = weekResult.CountOrders;
             }
@@ -156,7 +148,7 @@ namespace SmartStore.Services.Orders
             {
                 DateTime? startTime3 = _dateTimeHelper.ConvertToUtcTime(t3, timeZone);
                 DateTime? endTime3 = null;
-				var monthResult = GetOrderAverageReportLine(storeId, os, null, null, startTime3, endTime3, null);
+				var monthResult = GetOrderAverageReportLine(storeId, orderStatusId, null, null, startTime3, endTime3, null);
                 item.SumThisMonthOrders = monthResult.SumOrders;
                 item.CountThisMonthOrders = monthResult.CountOrders;
             }
@@ -166,14 +158,14 @@ namespace SmartStore.Services.Orders
             {
                 DateTime? startTime4 = _dateTimeHelper.ConvertToUtcTime(t4, timeZone);
                 DateTime? endTime4 = null;
-				var yearResult = GetOrderAverageReportLine(storeId, os, null, null, startTime4, endTime4, null);
+				var yearResult = GetOrderAverageReportLine(storeId, orderStatusId, null, null, startTime4, endTime4, null);
                 item.SumThisYearOrders = yearResult.SumOrders;
                 item.CountThisYearOrders = yearResult.CountOrders;
             }
             //all time
             DateTime? startTime5 = null;
             DateTime? endTime5 = null;
-			var allTimeResult = GetOrderAverageReportLine(storeId, os, null, null, startTime5, endTime5, null);
+			var allTimeResult = GetOrderAverageReportLine(storeId, orderStatusId, null, null, startTime5, endTime5, null);
             item.SumAllTimeOrders = allTimeResult.SumOrders;
             item.CountAllTimeOrders = allTimeResult.CountOrders;
 
@@ -360,44 +352,43 @@ namespace SmartStore.Services.Orders
 		/// <param name="storeId">Store identifier</param>
         /// <param name="startTimeUtc">Start date</param>
         /// <param name="endTimeUtc">End date</param>
-        /// <param name="os">Order status; null to load all records</param>
-        /// <param name="ps">Order payment status; null to load all records</param>
-        /// <param name="ss">Shipping status; null to load all records</param>
+		/// <param name="orderStatusIds">Filter by order status</param>
+		/// <param name="paymentStatusIds">Filter by payment status</param>
+		/// <param name="shippingStatusIds">Filter by shipping status</param>
         /// <param name="billingEmail">Billing email. Leave empty to load all records.</param>
         /// <returns>Result</returns>
-		public virtual decimal ProfitReport(int storeId, 
-			OrderStatus? os, PaymentStatus? ps, ShippingStatus? ss,
-			DateTime? startTimeUtc, DateTime? endTimeUtc,
-            string billingEmail)
+		public virtual decimal ProfitReport(int storeId, int[] orderStatusIds, int[] paymentStatusIds, int[] shippingStatusIds,
+			DateTime? startTimeUtc, DateTime? endTimeUtc, string billingEmail)
         {
-            int? orderStatusId = null;
-            if (os.HasValue)
-                orderStatusId = (int)os.Value;
-
-            int? paymentStatusId = null;
-            if (ps.HasValue)
-                paymentStatusId = (int)ps.Value;
-
-            int? shippingStatusId = null;
-            if (ss.HasValue)
-                shippingStatusId = (int)ss.Value;
             //We cannot use String.IsNullOrEmpty(billingEmail) in SQL Compact
             bool dontSearchEmail = String.IsNullOrEmpty(billingEmail);
-            var query = from orderItem in _orderItemRepository.Table
-                        join o in _orderRepository.Table on orderItem.OrderId equals o.Id
-						where (storeId == 0 || storeId == o.StoreId) &&
-							  (!startTimeUtc.HasValue || startTimeUtc.Value <= o.CreatedOnUtc) &&
-                              (!endTimeUtc.HasValue || endTimeUtc.Value >= o.CreatedOnUtc) &&
-                              (!orderStatusId.HasValue || orderStatusId == o.OrderStatusId) &&
-                              (!paymentStatusId.HasValue || paymentStatusId == o.PaymentStatusId) &&
-                              (!shippingStatusId.HasValue || shippingStatusId == o.ShippingStatusId) &&
-                              (!o.Deleted) &&
-                              (dontSearchEmail || (o.BillingAddress != null && !String.IsNullOrEmpty(o.BillingAddress.Email) && o.BillingAddress.Email.Contains(billingEmail)))
-                        select orderItem;
+
+			var orders =
+				from o in _orderRepository.Table
+				where (storeId == 0 || storeId == o.StoreId) &&
+					(!startTimeUtc.HasValue || startTimeUtc.Value <= o.CreatedOnUtc) &&
+					(!endTimeUtc.HasValue || endTimeUtc.Value >= o.CreatedOnUtc) &&
+					(!o.Deleted) &&
+					(dontSearchEmail || (o.BillingAddress != null && !String.IsNullOrEmpty(o.BillingAddress.Email) && o.BillingAddress.Email.Contains(billingEmail)))
+				select o;
+
+			if (orderStatusIds != null && orderStatusIds.Count() > 0)
+				orders = orders.Where(x => orderStatusIds.Contains(x.OrderStatusId));
+
+			if (paymentStatusIds != null && paymentStatusIds.Count() > 0)
+				orders = orders.Where(x => paymentStatusIds.Contains(x.PaymentStatusId));
+
+			if (shippingStatusIds != null && shippingStatusIds.Count() > 0)
+				orders = orders.Where(x => shippingStatusIds.Contains(x.ShippingStatusId));
+
+			var query =
+				from orderItem in _orderItemRepository.Table
+				join o in orders on orderItem.OrderId equals o.Id
+				select orderItem;
 
 			var productCost = Convert.ToDecimal(query.Sum(orderItem => (decimal?)orderItem.ProductCost * orderItem.Quantity));
 
-			var reportSummary = GetOrderAverageReportLine(storeId, os, ps, ss, startTimeUtc, endTimeUtc, billingEmail);
+			var reportSummary = GetOrderAverageReportLine(storeId, orderStatusIds, paymentStatusIds, shippingStatusIds, startTimeUtc, endTimeUtc, billingEmail);
 			var profit = reportSummary.SumOrders - reportSummary.SumTax - productCost;
 
             return profit;