+++ src/Presentation/SmartStore.Web.Framework/Extensions.cs
@@ -118,15 +118,20 @@ namespace SmartStore.Web.Framework
 
         public static GridBoundColumnBuilder<T> Centered<T>(this GridBoundColumnBuilder<T> columnBuilder) where T:class
         {
-            return columnBuilder.HtmlAttributes(new { align = "center" })
-                            .HeaderHtmlAttributes(new { style = "text-align:center;" });
+            return columnBuilder.HtmlAttributes(new { align = "center" }).HeaderHtmlAttributes(new { style = "text-align:center;" });
         }
-
         public static GridTemplateColumnBuilder<T> Centered<T>(this GridTemplateColumnBuilder<T> columnBuilder) where T : class
         {
-            return columnBuilder.HtmlAttributes(new { align = "center" })
-                            .HeaderHtmlAttributes(new { style = "text-align:center;" });
+            return columnBuilder.HtmlAttributes(new { align = "center" }).HeaderHtmlAttributes(new { style = "text-align:center;" });
         }
+		public static GridBoundColumnBuilder<T> RightAlign<T>(this GridBoundColumnBuilder<T> columnBuilder) where T : class
+		{
+			return columnBuilder.HtmlAttributes(new { style = "text-align:right;" }).HeaderHtmlAttributes(new { style = "text-align:right;" });
+		}
+		public static GridTemplateColumnBuilder<T> RightAlign<T>(this GridTemplateColumnBuilder<T> columnBuilder) where T : class
+		{
+			return columnBuilder.HtmlAttributes(new { style = "text-align:right;" }).HeaderHtmlAttributes(new { style = "text-align:right;" });
+		}
 
         public static SelectList ToSelectList<TEnum>(this TEnum enumObj, bool markCurrentAsSelected = true) where TEnum : struct
         {